fixed 39d980d from https://gitee.com/pcwlno1/hiviewdfx_hiview_lite/pulls/54 fix: BBOX兼容ARM平台
1、在核心框架中使用基础数据类型。
2、尽量使用用户层的函数接口。

Close #I422VS

+++ blackbox/blackbox_adapter.c
@@ -15,8 +15,8 @@
 
 /*-----------includes---------*/
 #include "blackbox.h"
+#include "blackbox_adapter.h"
 #include "blackbox_detector.h"
-#include "ohos_types.h"
 #include "ohos_init.h"
 
 /*----------local macroes-----------*/
@@ -51,7 +51,7 @@ WEAK int SystemModuleSaveLastLog(const char *logDir, struct ErrorInfo *info)
 }
 
 WEAK int FullWriteFile(const char *filePath, const char *buf,
-    size_t bufSize, int isAppend)
+    unsigned int bufSize, int isAppend)
 {
     (void)filePath;
     (void)buf;
